remove unused code and fix header
If a header already exists (e.g., from a prior LAS read), then it makes no sense
to force the scale to 1, etc. The problem is if the header is empty, i.e., the
data was originally read from another format. In this case, we will need to add
code to generate a new header, perhaps calculating scale, etc.